Replace and remove boilerplate icon, shrink other PNGs
`logo192.png` is referenced in the CSS for mobile so it is loaded but it was
the React logo, I replaced it with a 192x192 inferno one and optimized it but
it is bigger (more information).  To make up for this size increase (while we
are worried about initial page load), I shrunk the other PNG images with `optipng`
(no quality loss to my eyes).

The total image size of the images involved in this change was 265.19KB.  Now it is 195.5KB.