Change usage of zustand store to be more precise (#204)
Fetching everything from the store will cause unnecessary rerenders.  Doing it this way will be a good example for the future.  Have to be careful about selectors in Redux too.

Additionally, not all this scope is needed but this is the easiest to read.  `setTestSuites` is used on one location and could be moved into the first useEffect().  Until we get used to a global store, this might be the most familiar way to do it for now.