Transition test runner to vitest
The react-markdown plugin that will enable github flavored markdown
(which we want for table rendering to meet parity with the legacy app)
does not work with Jest because of ES Modules.

Vitest works with ES Modules and has less config and boilerplate than
Jest and babel.  We cannot move to Vite (the dev server and builder)
because of unknown issues (could be the target, could be our hanami setup)
but we **can** move to vitest all by itself.

Vitest is about 2x as fast on our app:
```
entire test suite: vitest:3.86s jest:7.39s
single test file:  vitest:2.13s jest:4.61s
```

The syntax is very similar to Jest, with only small modifications needed for
mocking App.tsx which is a bit of an oddball because it's the root
component.

I had to bump some eslint packages to current versions to get a plugin
to stop warning.  This caused some linting and typescript errors that
were `any` (tab index).  So I fixed those.