Fix markdown tables not rendering in messages (#248)
Github Flavored Markdown (GFM) has a tables syntax.  GFM is not in the
react-markdown library we are using.  Add the GFM plugin that makes table
rendering possible.  In addition, the plugin did not work at first with
the existing versions that were locked (react-markdown 6.0.2):

```
Uncaught TypeError: self.parser.lazy is undefined
```

So I bumped the versions to latest of react-markdown and the gfm plugin and
then it worked.