Fix pane scrolling bug when developing a testkit locally
The TestSession component has a left pane and a TestSuite right pane.
The bug on main is that they both scroll together.  This is because
a style is missing on this element `maxHeight`.  I believe this is
happening because of the javascript template literal.  I saw
inflection locally.  Essentially I built the gem off of main and
saw the bug, then I changed to a static string and the bug was fixed.
Then I changed this to use a function (essentially same behavior as the template literal)
and the bug was still fixed.

This was surprising.  I don't know why you can't
```
`calc(100vh - 64px - 56px - ${bannerHeight}px)`
```
When bannerHeight is just a variable but it doesn't seem to work.

My guess is maybe this code is loaded before the DOM fires.  In that case
having a binding function would fix this.  I bet the template literal errors or
becomes undefined and the build process strips out the style.