Fix UI bug where the TestListItem accordian would expand to the wrong tab
A useEffect that was concerned with setting the tab index when the tab
was empty of content was conflicting with another function which would
set the tab index to the clicked on "ProblemBadge".  In theory, the old
useEffect should not have affected the badge clicking but it clearly is.

Changing from the hook that fires on a data dependency to one that fires
on the accordian expansion click solves the problem.  I've added tests
to cover the previous situations.

I did not move the TabProps interface out of TestRunDetail.  I think
that's a good place for it but I am using the type in the parent element.