Replaced npm start command with direct webpack invocation
Because npm is not a process manager, it does not propagate SIGINTs to the scripts it starts
(webpack in this case) - thus, doing Ctrl+C on the Foreman process will kill the npm process
but not the webpack process.  Replacing the "npm run start" command with whatever is
listed under "start" in package.json bypasses this signal handling discrepancy.