Narrower error handling on validator and evaluator request
Add readable strings to help in debugging or error reports.  The network can fail anytime
and this will help narrow-in on the problem by having more precise error messages.

Add more error handling paths to where the validation request
happens.  There is another place where a similar request happens.
If this an acceptable idea I can change that there too.

This also should have an effect on the UI message because it
seems to use the String from the `raise`.